factor out stack test cases
