+++ src/test/java/org/jal/collections/stack/StackTest.java
@@ -6,13 +6,16 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.util.NoSuchElementException;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-public class ArrayStackTest {
+abstract class CommonTest {
+  abstract protected Stack<Integer> createStack();
+
   @DisplayName("peek() should be top")
   @Test
   public void testPushAndPeek() {
-    ArrayStack<Integer> stack = new ArrayStack<>();
+    Stack<Integer> stack = createStack();
     stack.push(42);
     stack.push(43);
 
@@ -22,7 +25,7 @@ public class ArrayStackTest {
   @DisplayName("pop() should be top")
   @Test
   public void testPushAndPop() {
-    ArrayStack<Integer> stack = new ArrayStack<>();
+    Stack<Integer> stack = this.createStack();
     stack.push(42);
     stack.push(43);
 
@@ -32,7 +35,7 @@ public class ArrayStackTest {
   @DisplayName("pop() should throw if empty")
   @Test
   public void testPopForEmptyStack() {
-    ArrayStack<Integer> stack = new ArrayStack<>();
+    Stack<Integer> stack = this.createStack();
 
     assertThrows(NoSuchElementException.class, () -> stack.pop());
   }
@@ -40,7 +43,7 @@ public class ArrayStackTest {
   @DisplayName("getSize() should get the size")
   @Test
   public void testGetSize() {
-    ArrayStack<Integer> stack = new ArrayStack<>();
+    Stack<Integer> stack = this.createStack();
     stack.push(42);
     stack.push(43);
 
@@ -50,7 +53,7 @@ public class ArrayStackTest {
   @DisplayName("isEmpty() should be true if empty")
   @Test
   public void testIsEmptyForEmptyStack() {
-    ArrayStack<Integer> stack = new ArrayStack<>();
+    Stack<Integer> stack = this.createStack();
 
     assertEquals(true, stack.isEmpty());
   }
@@ -58,9 +61,29 @@ public class ArrayStackTest {
   @DisplayName("isEmpty() should be false if nonempty")
   @Test
   public void testIsEmptyForNonemptyStack() {
-    ArrayStack<Integer> stack = new ArrayStack<>();
+    Stack<Integer> stack = this.createStack();
     stack.push(42);
 
     assertEquals(false, stack.isEmpty());
   }
+}
+
+public class StackTest {
+  @DisplayName("ArrayStack class")
+  @Nested
+  class ArrayStackTest extends CommonTest {
+    @Override
+    protected Stack<Integer> createStack() {
+      return new ArrayStack<Integer>();
+    }
+  };
+
+  @DisplayName("ListStack class")
+  @Nested
+  class ListStackTest extends CommonTest {
+    @Override
+    protected Stack<Integer> createStack() {
+      return new ListStack<Integer>();
+    }
+  };
 }
